# [ci] [python] reduce unnecessary data loading in tests