--- a/tests/python_package_test/test_engine.py
+++ b/tests/python_package_test/test_engine.py
@@ -7,6 +7,8 @@ import psutil
 import random
 import unittest
 
+from functools import lru_cache
+
 import lightgbm as lgb
 import numpy as np
 from scipy.sparse import csr_matrix, isspmatrix_csr, isspmatrix_csc
@@ -51,9 +53,26 @@ def categorize(continuous_x):
     return np.digitize(continuous_x, bins=np.arange(0, 1, 0.01))
 
 
+@lru_cache
+def _load_boston(**kwargs):
+    return load_boston(**kwargs)
+
+@lru_cache
+def _load_breast_cancer(**kwargs):
+    return load_breast_cancer(**kwargs)
+
+@lru_cache
+def _load_digits(**kwargs):
+    return load_digits(**kwargs)
+
+@lru_cache
+def _load_iris(**kwargs):
+    return load_iris(**kwargs)
+
+
 class TestEngine(unittest.TestCase):
     def test_binary(self):
-        X, y = load_breast_cancer(return_X_y=True)
+        X, y = _load_breast_cancer(return_X_y=True)
         X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42)
         params = {
             'objective': 'binary',
@@ -75,7 +94,7 @@ class TestEngine(unittest.TestCase):
         self.assertAlmostEqual(evals_result['valid_0']['binary_logloss'][-1], ret, places=5)
 
     def test_rf(self):
-        X, y = load_breast_cancer(return_X_y=True)
+        X, y = _load_breast_cancer(return_X_y=True)
         X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42)
         params = {
             'boosting_type': 'rf',
@@ -100,7 +119,7 @@ class TestEngine(unittest.TestCase):
         self.assertAlmostEqual(evals_result['valid_0']['binary_logloss'][-1], ret, places=5)
 
     def test_regression(self):
-        X, y = load_boston(return_X_y=True)
+        X, y = _load_boston(return_X_y=True)
         X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42)
         params = {
             'metric': 'l2',
@@ -377,7 +396,7 @@ class TestEngine(unittest.TestCase):
         self.assertAlmostEqual(evals_result['valid_0']['auc'][-1], ret, places=5)
 
     def test_multiclass(self):
-        X, y = load_digits(n_class=10, return_X_y=True)
+        X, y = _load_digits(n_class=10, return_X_y=True)
         X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42)
         params = {
             'objective': 'multiclass',
@@ -398,7 +417,7 @@ class TestEngine(unittest.TestCase):
         self.assertAlmostEqual(evals_result['valid_0']['multi_logloss'][-1], ret, places=5)
 
     def test_multiclass_rf(self):
-        X, y = load_digits(n_class=10, return_X_y=True)
+        X, y = _load_digits(n_class=10, return_X_y=True)
         X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42)
         params = {
             'boosting_type': 'rf',
@@ -426,7 +445,7 @@ class TestEngine(unittest.TestCase):
         self.assertAlmostEqual(evals_result['valid_0']['multi_logloss'][-1], ret, places=5)
 
     def test_multiclass_prediction_early_stopping(self):
-        X, y = load_digits(n_class=10, return_X_y=True)
+        X, y = _load_digits(n_class=10, return_X_y=True)
         X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42)
         params = {
             'objective': 'multiclass',
@@ -452,7 +471,7 @@ class TestEngine(unittest.TestCase):
         self.assertLess(ret, 0.2)
 
     def test_multi_class_error(self):
-        X, y = load_digits(n_class=10, return_X_y=True)
+        X, y = _load_digits(n_class=10, return_X_y=True)
         params = {'objective': 'multiclass', 'num_classes': 10, 'metric': 'multi_error',
                   'num_leaves': 4, 'verbose': -1}
         lgb_data = lgb.Dataset(X, label=y)
@@ -497,7 +516,7 @@ class TestEngine(unittest.TestCase):
 
     def test_auc_mu(self):
         # should give same result as binary auc for 2 classes
-        X, y = load_digits(n_class=10, return_X_y=True)
+        X, y = _load_digits(n_class=10, return_X_y=True)
         y_new = np.zeros((len(y)))
         y_new[y != 0] = 1
         lgb_X = lgb.Dataset(X, label=y_new)
@@ -575,7 +594,7 @@ class TestEngine(unittest.TestCase):
         self.assertNotEqual(results_weight['training']['auc_mu'][-1], results_no_weight['training']['auc_mu'][-1])
 
     def test_early_stopping(self):
-        X, y = load_breast_cancer(return_X_y=True)
+        X, y = _load_breast_cancer(return_X_y=True)
         params = {
             'objective': 'binary',
             'metric': 'binary_logloss',
@@ -607,7 +626,7 @@ class TestEngine(unittest.TestCase):
         self.assertIn('binary_logloss', gbm.best_score[valid_set_name])
 
     def test_continue_train(self):
-        X, y = load_boston(return_X_y=True)
+        X, y = _load_boston(return_X_y=True)
         X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42)
         params = {
             'objective': 'regression',
@@ -635,7 +654,7 @@ class TestEngine(unittest.TestCase):
         os.remove(model_name)
 
     def test_continue_train_reused_dataset(self):
-        X, y = load_boston(return_X_y=True)
+        X, y = _load_boston(return_X_y=True)
         params = {
             'objective': 'regression',
             'verbose': -1
@@ -648,7 +667,7 @@ class TestEngine(unittest.TestCase):
         self.assertEqual(gbm.current_iteration(), 20)
 
     def test_continue_train_dart(self):
-        X, y = load_boston(return_X_y=True)
+        X, y = _load_boston(return_X_y=True)
         X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42)
         params = {
             'boosting_type': 'dart',
@@ -671,7 +690,7 @@ class TestEngine(unittest.TestCase):
         self.assertAlmostEqual(evals_result['valid_0']['l1'][-1], ret, places=5)
 
     def test_continue_train_multiclass(self):
-        X, y = load_iris(return_X_y=True)
+        X, y = _load_iris(return_X_y=True)
         X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42)
         params = {
             'objective': 'multiclass',
@@ -694,7 +713,7 @@ class TestEngine(unittest.TestCase):
         self.assertAlmostEqual(evals_result['valid_0']['multi_logloss'][-1], ret, places=5)
 
     def test_cv(self):
-        X_train, y_train = load_boston(return_X_y=True)
+        X_train, y_train = _load_boston(return_X_y=True)
         params = {'verbose': -1}
         lgb_train = lgb.Dataset(X_train, y_train)
         # shuffle = False, override metric in params
@@ -753,7 +772,7 @@ class TestEngine(unittest.TestCase):
         np.testing.assert_allclose(cv_res_lambda['ndcg@3-mean'], cv_res_lambda_obj['ndcg@3-mean'])
 
     def test_cvbooster(self):
-        X, y = load_breast_cancer(return_X_y=True)
+        X, y = _load_breast_cancer(return_X_y=True)
         X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42)
         params = {
             'objective': 'binary',
@@ -797,7 +816,7 @@ class TestEngine(unittest.TestCase):
         self.assertLess(ret, 0.15)
 
     def test_feature_name(self):
-        X_train, y_train = load_boston(return_X_y=True)
+        X_train, y_train = _load_boston(return_X_y=True)
         params = {'verbose': -1}
         lgb_train = lgb.Dataset(X_train, y_train)
         feature_names = ['f_' + str(i) for i in range(X_train.shape[-1])]
@@ -825,7 +844,7 @@ class TestEngine(unittest.TestCase):
 
     def test_save_load_copy_pickle(self):
         def train_and_predict(init_model=None, return_model=False):
-            X, y = load_boston(return_X_y=True)
+            X, y = _load_boston(return_X_y=True)
             X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42)
             params = {
                 'objective': 'regression',
@@ -989,7 +1008,7 @@ class TestEngine(unittest.TestCase):
         self.assertEqual(len(evals_result['valid_1']['rmse']), 20)
 
     def test_contribs(self):
-        X, y = load_breast_cancer(return_X_y=True)
+        X, y = _load_breast_cancer(return_X_y=True)
         X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42)
         params = {
             'objective': 'binary',
@@ -1372,7 +1391,7 @@ class TestEngine(unittest.TestCase):
         np.random.seed()  # reset seed
 
     def test_refit(self):
-        X, y = load_breast_cancer(return_X_y=True)
+        X, y = _load_breast_cancer(return_X_y=True)
         X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42)
         params = {
             'objective': 'binary',
@@ -1388,7 +1407,7 @@ class TestEngine(unittest.TestCase):
         self.assertGreater(err_pred, new_err_pred)
 
     def test_mape_rf(self):
-        X, y = load_boston(return_X_y=True)
+        X, y = _load_boston(return_X_y=True)
         params = {
             'boosting_type': 'rf',
             'objective': 'mape',
@@ -1405,7 +1424,7 @@ class TestEngine(unittest.TestCase):
         self.assertGreater(pred_mean, 20)
 
     def test_mape_dart(self):
-        X, y = load_boston(return_X_y=True)
+        X, y = _load_boston(return_X_y=True)
         params = {
             'boosting_type': 'dart',
             'objective': 'mape',
@@ -1484,7 +1503,7 @@ class TestEngine(unittest.TestCase):
             params['num_class'] = 4
             return dtrain, dtest, params
 
-        X, y = load_iris(return_X_y=True)
+        X, y = _load_iris(return_X_y=True)
         dataset = lgb.Dataset(X, y, free_raw_data=False)
         params = {'objective': 'multiclass', 'num_class': 3, 'verbose': -1}
         results = lgb.cv(params, dataset, num_boost_round=10, fpreproc=preprocess_data)
@@ -1492,7 +1511,7 @@ class TestEngine(unittest.TestCase):
         self.assertEqual(len(results['multi_logloss-mean']), 10)
 
     def test_metrics(self):
-        X, y = load_digits(n_class=2, return_X_y=True)
+        X, y = _load_digits(n_class=2, return_X_y=True)
         X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42)
         lgb_train = lgb.Dataset(X_train, y_train, silent=True)
         lgb_valid = lgb.Dataset(X_test, y_test, reference=lgb_train, silent=True)
@@ -1798,7 +1817,7 @@ class TestEngine(unittest.TestCase):
         self.assertEqual(len(evals_result), 1)
         self.assertIn('error', evals_result['valid_0'])
 
-        X, y = load_digits(n_class=3, return_X_y=True)
+        X, y = _load_digits(n_class=3, return_X_y=True)
         lgb_train = lgb.Dataset(X, y, silent=True)
 
         obj_multi_aliases = ['multiclass', 'softmax', 'multiclassova', 'multiclass_ova', 'ova', 'ovr']
@@ -1866,7 +1885,7 @@ class TestEngine(unittest.TestCase):
                           params_class_3_verbose, metrics='binary_error', fobj=dummy_obj)
 
     def test_multiple_feval_train(self):
-        X, y = load_breast_cancer(return_X_y=True)
+        X, y = _load_breast_cancer(return_X_y=True)
 
         params = {'verbose': -1, 'objective': 'binary', 'metric': 'binary_logloss'}
 
@@ -1889,7 +1908,7 @@ class TestEngine(unittest.TestCase):
         self.assertIn('decreasing_metric', evals_result['valid_0'])
 
     def test_multiple_feval_cv(self):
-        X, y = load_breast_cancer(return_X_y=True)
+        X, y = _load_breast_cancer(return_X_y=True)
 
         params = {'verbose': -1, 'objective': 'binary', 'metric': 'binary_logloss'}
 
@@ -1912,7 +1931,7 @@ class TestEngine(unittest.TestCase):
 
     @unittest.skipIf(psutil.virtual_memory().available / 1024 / 1024 / 1024 < 3, 'not enough RAM')
     def test_model_size(self):
-        X, y = load_boston(return_X_y=True)
+        X, y = _load_boston(return_X_y=True)
         data = lgb.Dataset(X, y)
         bst = lgb.train({'verbose': -1}, data, num_boost_round=2)
         y_pred = bst.predict(X)
@@ -1938,7 +1957,7 @@ class TestEngine(unittest.TestCase):
             self.skipTest('not enough RAM')
 
     def test_get_split_value_histogram(self):
-        X, y = load_boston(return_X_y=True)
+        X, y = _load_boston(return_X_y=True)
         lgb_train = lgb.Dataset(X, y, categorical_feature=[2])
         gbm = lgb.train({'verbose': -1}, lgb_train, num_boost_round=20)
         # test XGBoost-style return value
@@ -2048,7 +2067,7 @@ class TestEngine(unittest.TestCase):
                          eval_train_metric=eval_train_metric)
             self.assertEqual(assumed_iteration, len(ret[list(ret.keys())[0]]))
 
-        X, y = load_boston(return_X_y=True)
+        X, y = _load_boston(return_X_y=True)
         X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
         X_test1, X_test2, y_test1, y_test2 = train_test_split(X_test, y_test, test_size=0.5, random_state=73)
         lgb_train = lgb.Dataset(X_train, y_train)
@@ -2126,7 +2145,7 @@ class TestEngine(unittest.TestCase):
                                                                            decreasing_metric(preds, train_data)])
 
     def test_node_level_subcol(self):
-        X, y = load_breast_cancer(return_X_y=True)
+        X, y = _load_breast_cancer(return_X_y=True)
         X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.1, random_state=42)
         params = {
             'objective': 'binary',
@@ -2307,7 +2326,7 @@ class TestEngine(unittest.TestCase):
 
     def test_extra_trees(self):
         # check extra trees increases regularization
-        X, y = load_boston(return_X_y=True)
+        X, y = _load_boston(return_X_y=True)
         lgb_x = lgb.Dataset(X, label=y)
         params = {'objective': 'regression',
                   'num_leaves': 32,
@@ -2325,7 +2344,7 @@ class TestEngine(unittest.TestCase):
 
     def test_path_smoothing(self):
         # check path smoothing increases regularization
-        X, y = load_boston(return_X_y=True)
+        X, y = _load_boston(return_X_y=True)
         lgb_x = lgb.Dataset(X, label=y)
         params = {'objective': 'regression',
                   'num_leaves': 32,
@@ -2347,7 +2366,7 @@ class TestEngine(unittest.TestCase):
             cols = ['Column_' + str(i) for i in range(X.shape[1])]
             return [impcts_dict.get(col, 0.) for col in cols]
 
-        X, y = load_breast_cancer(return_X_y=True)
+        X, y = _load_breast_cancer(return_X_y=True)
         data = lgb.Dataset(X, label=y)
         num_trees = 10
         bst = lgb.train({"objective": "binary", "verbose": -1}, data, num_trees)
@@ -2392,7 +2411,7 @@ class TestEngine(unittest.TestCase):
             self.assertIsNone(tree_df.loc[0, col])
 
     def test_interaction_constraints(self):
-        X, y = load_boston(return_X_y=True)
+        X, y = _load_boston(return_X_y=True)
         num_features = X.shape[1]
         train_data = lgb.Dataset(X, label=y)
         # check that constraint containing all features is equivalent to no constraint
@@ -2469,7 +2488,7 @@ class TestEngine(unittest.TestCase):
             np.testing.assert_allclose(pred4, pred6)
 
         # test for regression
-        X, y = load_boston(return_X_y=True)
+        X, y = _load_boston(return_X_y=True)
         params = {
             'objective': 'regression',
             'verbose': -1,
@@ -2482,7 +2501,7 @@ class TestEngine(unittest.TestCase):
         inner_test(X, y, params, early_stopping_rounds=None)
 
         # test for multi-class
-        X, y = load_iris(return_X_y=True)
+        X, y = _load_iris(return_X_y=True)
         params = {
             'objective': 'multiclass',
             'metric': 'multi_logloss',
@@ -2496,7 +2515,7 @@ class TestEngine(unittest.TestCase):
         inner_test(X, y, params, early_stopping_rounds=None)
 
         # test for binary
-        X, y = load_breast_cancer(return_X_y=True)
+        X, y = _load_breast_cancer(return_X_y=True)
         params = {
             'objective': 'binary',
             'metric': 'binary_logloss',