--- a/tests/python_package_test/utils.py
+++ b/tests/python_package_test/utils.py
@@ -2,6 +2,7 @@
 try:
     from functools import lru_cache
 except ImportError:
+    import warnings
     warnings.warn("Could not import functools.lru_cache", RuntimeWarning)
 
     def lru_cache(user_function, maxsize=None):