I spent some time this weekend learning [`snakeviz`](https://jiffyclub.github.io/snakeviz/), a Python profiler. I tried running it over the Python package tests here, and I think I found an opportunity to cut the test run time.


I realized we're spending time loading and reloading the same datasets.  This PR proposes reducing those calls by caching the results of `sklearn.datasets.load_*` calls. The datasets are really small so I don't think holding them all in memory will cause an issue.

**UPDATE:** Ok from some experiments, the speedup from this seems really small on my laptop. But might be larger in CI environments, where we know that disk I/O is a lot slower (https://github.com/microsoft/LightGBM/pull/2965#issuecomment-624421092).

## Results

These results were obtained on my laptop.  The blockquote results come from one run of each test. Then I ran them all again 10 times to get a better estimate of the mean time.

### `test_basic.py`

**before (mean = 1.33s)**

> 58216 function calls (57604 primitive calls) in 0.462 seconds
    ==== 10 passed, 3 warnings in 2.18s ====

* trials (seconds):  [1.37, 1.40, 1.34, 1.37, 1.31, 1.30, 1.31, 1.31, 1.31, 1.31]

**after (mean = 1.36s)**

> 53470 function calls (52858 primitive calls) in 0.396 seconds
    ==== 10 passed, 3 warnings in 1.36s ====

* trials (seconds):  [1.39, 1.39, 1.40, 1.36, 1.35, 1.41, 1.40, 1.29, 1.31, 1.35]

### `test_engine.py`

**before (mean = 12.78s)**

> 2749043 function calls (2651956 primitive calls) in 9.456 seconds
    ==== 63 passed, 2 skipped, 20 warnings in 13.69s ====

* trials (seconds): [14.16, 13.13, 12.85, 12.66, 12.34, 12.60, 13.17, 12.98, 12.40, 11.50]

**after (mean = 11.84s)**

> 2749043 function calls (2651956 primitive calls) in 9.456 seconds
    ==== 63 passed, 2 skipped, 20 warnings in 11.29s ====

* trials (seconds): [12.18, 10.71, 11.09, 10.88, 10.77, 12.64, 12.47, 14.90, 11.62, 11.14]

### `test_sklearn.py`

**before(mean = 9.68s)**

> 4292703 function calls (4172793 primitive calls) in 8.044 seconds
    ==== 34 passed, 15 warnings in 11.07s ====

* trials (seconds): [10.17, 9.52, 9.34, 9.58, 9.22, 9.34, 9.75, 10.21, 9.67, 9.99]

**after (mean = 9.50s)**

> 2945200 function calls (2843392 primitive calls) in 8.867 seconds
    ==== 34 passed, 15 warnings in 9.26s ====

* trials (seconds): [9.67, 8.64, 9.32, 8.71, 8.73, 9.29, 8.97, 12.73, 9.48, 9.44]

## How to reproduce these tests

```shell
# install lightgbm
pushd python-package
python setup.py install
popd

# install dependencies
pip install snakeviz pytest-profiling

# profile tests
pytest --profile tests/python_package_test/test_basic.py
pytest --profile tests/python_package_test/test_engine.py
pytest --profile tests/python_package_test/test_sklearn.py

# (optional) visualize profiling data
snakeviz prof/combined.prof
```