@jameslamb 2020-10-26
interesting, I see this in some tests:

```text
../../../../miniconda/envs/test-env/lib/python3.6/functools.py:477: in lru_cache
    raise TypeError('Expected maxsize to be an integer or None')
E   TypeError: Expected maxsize to be an integer or None
```

Maybe there was not a default value in older versions of Python? Because today it has a default of 128: https://docs.python.org/3/library/functools.html#functools.lru_cache

Anyway, I'm going to try switching these to just `@cache`. Since they're just tests that we completely control, not user code, I think it's ok to have an unbounded cache. And that should be faster. From https://docs.python.org/3/library/functools.html#functools.cache:

> Returns the same as lru_cache(maxsize=None), creating a thin wrapper around a dictionary lookup for the function arguments. Because it never needs to evict old values, this is smaller and faster than lru_cache() with a size limit.