@StrikerRUS 2020-10-26
@jameslamb I think we can have `maxsize` equals to `2` for functions with only one bool argument `return_X_y=True/False` and 32 for `load_digits`. `None` doesn't look as a good default value:
https://github.com/python/cpython/blob/920cb647ba23feab7987d0dac1bd63bfc2ffc4c0/Lib/functools.py#L549-L562