@jameslamb 2020-10-26
@StrikerRUS can you explain why you think `None` is a bad default?

From that code snippet, it looks like it would be faster than setting a `maxsize` (since there is no code about evicting things from the cache).

I understand why it would be bad in user code, since the cache can grow without limit, but for these unit tests where we completely control the set of unique combinations and know it to be small, I think we should have a preference for the faster option.