@StrikerRUS 2020-10-26
@jameslamb 
`maxsize=None` means no LRU feature. I thought you want to use it.