@jameslamb 2020-10-26
> @jameslamb
> `maxsize=None` means no LRU feature. I thought you want to use it.

oh I see. No I really just cared more about the caching than using Least Recently Used (LRU), since we have so few variations of kwargs for each dataset.

I did forget though that `functools.lru_cache` isn't available in Python 2.7. Will push something to fix that :grimacing: