@StrikerRUS 2020-10-26
@jameslamb Ah OK! I misunderstood you then.